shiina

+++ ssm01/src/main/java/com/itcast/mapper/UserMapper.java
@@ -21,6 +21,6 @@ public interface UserMapper {
     List<User> selectUsers();
 
     @Select("insert into user values(#{username},#{money})")
-    int insertUser(@Param("username") String username,@Param("money") String money);
+    long insertUser(@Param("username") String username,@Param("money") String money);
 
 }
